Add file_copy()

+++ src/core/config.cpp
@@ -211,8 +211,8 @@ void parse(const config_t &config,
 
   mat.resize(rows, cols);
   size_t index = 0;
-  for (int i = 0; i < rows; i++) {
-    for (int j = 0; j < cols; j++) {
+  for (size_t i = 0; i < rows; i++) {
+    for (size_t j = 0; j < cols; j++) {
       mat(i, j) = node["data"][index].as<double>();
       index++;
     }
